requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pillow==9.5.0
requests==2.31.0
mysql-connector-python==8.0.33
